+++ packages/shared-base/src/utils/compose.ts
@@ -1,0 +1,6 @@
+export function compose<T>(...args: [...composer: Array<((arg: T) => T) | null | false>, init: T]) {
+    if (args.length === 0) throw new TypeError()
+    const last = args.pop() as T
+    // eslint-disable-next-line unicorn/no-array-reduce
+    return (args as Array<((arg: T) => T) | null>).filter(Boolean).reduceRight((prev, fn) => fn!(prev), last)
+}
