+++ packages/shared-base/src/utils/isDashboardPage.ts
@@ -1,7 +1,3 @@
-export function isPopupPage() {
-    return location.protocol.includes('extension') && location.href.includes('popups.html')
-}
-
 export function isDashboardPage() {
     return location.protocol.includes('extension') && location.href.includes('dashboard.html')
 }
