+++ packages/shared-base/src/utils/isPopupPage.ts
@@ -1,0 +1,3 @@
+export function isPopupPage() {
+    return location.protocol.includes('extension') && location.href.includes('popups.html')
+}
